Move ConfirmationToken related classes to new directory
