+++ src/main/java/com/koliving/api/registeration/token/ConfirmationTokenRepository.java
@@ -1,8 +1,7 @@
-package com.koliving.api.email;
+package com.koliving.api.registeration.token;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
